refactor: LpCommentSkeleton 위치 이동, LP import순서
